+++ packages/program/src/index.ts
@@ -5,3 +5,4 @@ export { MultiCallInvocationScope } from './functions/multicall-scope';
 export { default as Contract } from './contract';
 export { ScriptRequest } from './script-request';
 export { InstructionSet } from './instruction-set';
+export * from './response';
